requirements.txt
asgiref==3.11.1
dj-database-url==3.1.2
Django==6.0.2
django-cors-headers==4.9.0
djangorestframework==3.16.1
gunicorn==25.1.0
packaging==26.0
pillow==12.1.1
sqlparse==0.5.5
whitenoise==6.11.0
